requirements.txt
Appium-Python-Client==0.42
selenium==3.141.0
urllib3==1.25.2
